server/requirements.txt
Flask
mysql-connector-python
bcrypt
pyjwt